monospaced optimization fix, missing [global] in abbrevs
git-svn-id: svn+ssh://svn.code.sf.net/p/jedit/svn/jEdit/trunk@3259 6b1eeb88-9816-0410-afa2-b43733a0f04e

+++ org/gjt/sp/jedit/textarea/TextAreaPainter.java
@@ -877,7 +877,7 @@ public class TextAreaPainter extends JComponent implements TabExpander
 				.getWidth();
 			double width2 = str.length() * maxWidth;
 			if(minWidth == maxWidth
-				&& width1 == width2)
+				&& (int)width1 == (int)width2)
 			{
 				Log.log(Log.DEBUG,this,"Using monospaced font optimization: " + font);
 				returnValue = new Integer(maxWidth);
